Always use non-legacy file chooser on Android Q and higher
The legacy chooser requires filesystem access to the external files
directory which is deprecated on Q and higher.  The non-legacy chooser
uses the storage access framework which should work on all versions.

Update the preferences to not allow the option on Q and higher.

Update layout for showing storage list during default file choosing to
wrap the files available.  Add a button to select no default file.  Do
not show the file open button.

Update unit tests to work with and without availability of the external
files.  Skip tests for the legacy file chooser.  Do not require a test
.psafe3 file being previously installed.