Fix de-selection of non-visible entry
If the view holder is not visible but not yet recycled, its selection
wasn't being updated when a new item was selected.  The
StableIdKeyProvider issues a warning about the key no longer resolving
to a position and doesn't inform the adapter.

In the selection observer, check for the same condition and schedule an
update to the adapter's data set to force a redraw.