Fix recent files disappearing
If the Google Drive app is closed (e.g. force closed), the next time
PasswdSafe opens, the query for the recent file URI will fail.  Likely,
the app hasn't been restored to provide a response to the query (though
subsequent queries can succeed after some time).

Remove the check for the cursor existing, and always take the
permission.   Perform the query so that the URI resolves during the
file open resolve.  The file is only removed if the query or permission
take fails.