Move open progress bar to left of Open button
The progress bar is moved from the right of the Yubikey progress text to
next to the Open button.  The Yubikey text is being expanded to include
more text for use of USB or NFC YubiKeys.

Cleanup warnings about excessively wide input and check fields.  Cleanup
warnings about too short check fields.

Remove the progress barrier as the Open button is used to stack the
components.

Move the help text field below the Open button where it is located
visually.