sync: onedrive: Support quick XOR hash of file contents
The OneDrive service is deprecating the use of SHA hashes for file contents in
favor of their quick XOR hash.  Use the SHA1 hash if still present for
compatibility and otherwise the quick XOR hash.

Existing sync files may return the SHA1 hash until they are updated resulting in
a change to the quick XOR hash style.