sync: onedrive: Update Proguard rules for library update
Incorporate rules from msgraph-sdk-java proguard-rules.txt in docs
directory.  The GSON rules appear to be included already by the GSON
library.  The IJsonBackedObject rule is not included as it appears to
have been used by a previous version of the msgraph library.

Add a specific rule for DriveItemCollectionResponse to work with
PageIterator on a minified build.

Add extra rules needed by a manual build with gradle.