passwdsafe: GH-23 Add option to display on external monitor
If external displays are supported, add a preference to allow the main password
activity to display on an external display.  By default, the activity is marked
that it shouldn't appear in the recent activities list.  An external display is
not necessarily secure, so an option is added to allow display.